backend/requirements.txt
asgiref==3.3.2
Django==3.2
django-filter==22.1
djangorestframework==3.12.4
gunicorn==20.1.0
djoser
pillow
psycopg2-binary==2.9.3
python-dotenv
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2020.1
sqlparse==0.3.1
requests==2.26.0
drf-base64==2.0
PyYAML==6.0